sysuser: rename api.js to cloud.js
